[Refectoring]router & app.tsx 구조 children활용하여 RootLayout.tsx 생성
