+++ src/App.tsx
@@ -1,29 +1,26 @@
-import { BrowserRouter } from "react-router-dom";
-import TransPages from "./pages/TransPages";
-import Header from "./components/organism/header/Header";
-import { useMain } from "./hooks/useMain";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import "./styles/fonts.css";
+import { Navigate, Route, Routes, useLocation } from "react-router-dom";
+import MainPage from "./pages/MainPage";
+import AboutMePage from "./pages/AboutMePage";
+import ProjectPage from "./pages/ProjectPage";
+import ImagePage from "./pages/ImagePage";
+import ThreePage from "./pages/ThreePage";
+import VideoPage from "./pages/VideoPage";
+import StockPage from "./pages/StockPage";
 
 function App() {
-  const { isActive, toggleActive } = useMain();
-  const queryClient = new QueryClient({
-    defaultOptions: {
-      queries: {
-        refetchOnWindowFocus: false,
-      },
-    },
-  });
+  const location = useLocation();
+
   return (
-    <BrowserRouter>
-      <div onClick={(e) => isActive && toggleActive(e)}>
-        <Header isActive={isActive} toggleActive={toggleActive} />
-        <QueryClientProvider client={queryClient}>
-          <TransPages />
-        </QueryClientProvider>
-      </div>
-    </BrowserRouter>
+    <Routes location={location}>
+      <Route path="/" element={<MainPage />} />
+      <Route path="me" element={<AboutMePage />} />
+      <Route path="project" element={<ProjectPage />} />
+      <Route path="project/image" element={<ImagePage />} />
+      <Route path="project/three" element={<ThreePage />} />
+      <Route path="project/video" element={<VideoPage />} />
+      <Route path="project/stock" element={<StockPage />} />
+      <Route path="*" element={<Navigate to="/" />} />
+    </Routes>
   );
 }
-
 export default App;
