--- src/pages/index.ts
@@ -1,19 +1,0 @@
-// import { lazy } from "react";
-
-export { default as MainPage } from "../pages/MainPage";
-export { default as VideoPage } from "../pages/VideoPage";
-export { default as StockPage } from "./StockPage";
-export { default as AboutMePage } from "../pages/AboutMePage";
-export { default as ProjectPage } from "../pages/ProjectPage";
-export { default as ThreePage } from "../pages/ThreePage";
-export { default as ImagePage } from "../pages/ImagePage";
-
-// const MainPage = lazy(() => import("../pages/MainPage"))
-// const VideoPage = lazy(() => import("../pages/VideoPage"))
-// const StockPage = lazy(() => import("../pages/StockPage"))
-// const AboutMePage = lazy(() => import("../pages/AboutMePage"))
-// const ProjectPage = lazy(() => import("../pages/ProjectPage"))
-// const ThreePage = lazy(() => import("../pages/ThreePage"))
-// const ImagePage = lazy(() => import("../pages/ImagePage"))
-
-// export {MainPage, VideoPage, StockPage, AboutMePage, ProjectPage, ThreePage, ImagePage}
