+++ tsconfig.json
@@ -8,7 +8,12 @@
     "baseUrl": ".",
     "paths": {
       "@/*": ["src/*"],
-      "@/components": ["src/components/*"]
+      "@/assets": ["src/assets/*"],
+      "@/components": ["src/components/*"],
+      "@/hooks": ["src/hooks/*"],
+      "@/pages": ["src/pages/*"],
+      "@/styles": ["src/styles/*"],
+      "@/types": ["src/types/*"]
     },
 
     /* Bundler mode */
